Dodo: Update TravisCI to do a dry-run
Problem

A complete build does not finish in under the TravisCI time
limit and also runs OOM currently.

Solution

We still want to exercise the script in some manner as a
test, so change to a dry-run which will clone repos and echo
the commands but not run them.

RB_ID=854822
TBR=true