dodo: Clean resolution cache before running sbt update
Problem/Solution

As I understand it today (and this evolves), the intent of
cleaning the caches is to force resolution to pick up the locally
built dependency artifacts for a project. Running this *after*
running `./sbt update` does not make sense. We should clean the
resolution cache first then run resolution via `update`.

Differential Revision: https://phabricator.twitter.biz/D220397