dodo: Don't add scala version to scrooge-sbt-plugin cross-publishing
Problem

When passing the `--scala-version` argument it looks like Dodo is adding
it to the `scrooge-sbt-plugin` cross publishing commands for sbt 0.13 and
and sbt 1.x. It looks like setting a scala version takes precedent over
the `^^` commands and the plugin ends up published incorrectly for sbt
1.x when given a scala version that is not 2.12.

Solution

Much like we skip applying the passed version if the commands is detected
to already have a version specified, we also additionally skip if it is
detected that there is an sbt cross version specified since this implies
as specific scala version.

Differential Revision: https://phabricator.twitter.biz/D222005