dodo: Always publish scrooge-generator for Scala 2.12
Problem

Much like we always publish the `scrooge-generator` in Scala 2.10
because it is necessary for the `scrooge-sbt-plugin` with sbt 0.13
we also need to always publish the `scrooge-generator` in Scala
2.12 because it is necessary for the `scrooge-sbt-plugin` with
sbt 1.x. This is a problem when we pass a specific Scala version
to dodo, we need to ensure that we publish the `scrooge-generator`
with these Scala versions, regardless.

Solution

Add a command to the `SCROOGE_CORE` commands to ensure we always
attempt to publish the generator with Scala 2.12 such that it is
available to be resolved by the scrooge-sbt-plugin` with sbt 1.x.

TBR=true

Differential Revision: https://phabricator.twitter.biz/D222272