dodo: Cross publish scrooge-generator
Problem/Solution

Instead of explicitly publishing the scrooge-generator locally for
both 2.10.7 and 2.12.8 we can just cross publish locally for all
versions. This allows us to not have to maintain updating the
scala versions here and instead we can stick to the versions
specified in the scrooge/build.sbt. This comes at the cost of
potentially cross-publishing more versions than necessary but it
worth it to reduce the surface area of Scala version changes.

Differential Revision: https://phabricator.twitter.biz/D446930