Dodo: Use master HEAD timestamp for shallow-since depth
Problem

We'd like to ensure that the fetch depth for getting the
develop branch for a repository is always deep enough.

Solution

Update to use `--shallow-since` set to the timestamp of
the current HEAD in the master branch. Move this logic
into the `bin/publish` script and fix up the individual
project releases to make sure the merging of develop
branch into master for upstreams happens correctly.

Differential Revision: https://phabricator.twitter.biz/D484122