dodo: Do not build the scrooge-generator multiple times
Problem

`scrooge-generator` is listed in the published projects which is
aggregated into the `scroogePublishLocal` and is thus published
when we execute `scrooge-publish-local/publishLocal`. However, we
then publish the scrooge-generator again seperately which is
uneccessary. We used to publish the generator specifically in Scala
2.10 here but things were changed around and we thus should not
need to publish the generator separately.

Solution

Don't attempt to publish the `scrooge-generator` twice. Fix up
the `clean_ivy` task.

Differential Revision: https://phabricator.twitter.biz/D490864