dodo: Exclude Finatra examples
Problem/Solution

The finatra/examples are currently being incorrectly included
during publishing of the Finatra project. One of the examples
fails in the `doc` phase because of incorrectly specified
Javadocs (they are written as Scaladocs), causing publishing
of the project to fail.

Update to remove the finatra/examples from the root of the
finatra/build.sbt when publishing.

Differential Revision: https://phabricator.twitter.biz/D491006