dodo: update GitHub Action gpg config for gnupg2 support
Problem/Solution

We receive `gpg: gpg-agent is not available in this session`
messages in our SBT builds due to using an older version
of `gnupg` in our container and a mismatched newer version
of the `gpg-agent` process. Let's modify our config to be
forward compatible so that we can eventually update our
version of `gnupg` to match our `gpg-agent` in our
GitHub Actions container.

Differential Revision: https://phabricator.twitter.biz/D492509