dodo: Fix up Scrooge release steps
Problem(s)/Solutions

The publish script always needs to be passed `--dry-run` otherwise
it errors, thus we add that back to the `release-snapshots.yml`
workflow. Additionally we add a way to test the publish non-destructively
locally. Lastly, the `scrooge` project commands are out of order
and we should build the generator before the linter and other projects.
Since it has no other dependencies, we do it first.

Differential Revision: https://phabricator.twitter.biz/D496689