dodo: Split Adding and Commiting Files for Site Publish
Problem

In the case of running the site publish action for our libraries,
we would up with empty repos for scrooge, twitter-server, and finatra.
I believe it's due to our use of `git rm -fr .` and then `git commit
-am` later. When we remove the files, we remove them from the index, and
`git commit -am` does not add them back.

Solution

Switch to explicitly calling `git add .` and splitting that out from the
`git commit` call so that files that were 'removed' and 'added back' are
properly added back.

Differential Revision: https://phabricator.twitter.biz/D506473