dodo: Add --ignore-unmatch to git rm Commands
Problem

If `git rm` fails to remove any files, it is seen as a fatal
error with an exception message of 'pathspec _ did not match
any files'. If our site publishing scripts are trying to recover
from a scenario where the branch is empty, this failure will
be treated as fatal and not allow us to continue to fix the
problem.

Solution

Add the `--ignore-unmatch` option to `git rm` which will 'Exit
with a zero status even if no files matches.', which in the case
where we're trying to recover from an empty repo is what we want.

Differential Revision: https://phabricator.twitter.biz/D506538