dodo: Use Custom Shell for Site Publish Github Action
Problem

The publishing of util docs is now failing, but it's not
clear as to where. GitHub doesn't print out the command it's
running by default when specifying the shell as 'bash'.

Solution

Let's add a few additional options to the default shell used
for this action to aid in debugging.

Differential Revision: https://phabricator.twitter.biz/D506661