dodo: Remove Scala Version Workaround for Site Publishing
Problem / Solution

A workaround was added for release 20.7.0 so that it used Scala version
2.12.8 for publishing our documentation instead of Scala version 2.12.11
(which was used to compile the release). This was because unidoc
documentation generation failed when using Scala version 2.12.11, but
not 2.12.8. Since that time, the code which was causing the failure has
been updated and should no longer cause it to fail. So the workaround
should not be removed.

JIRA Issues: CSL-9977

Differential Revision: https://phabricator.twitter.biz/D520323