dodo: Skip snapshot when version is not '-SNAPSHOT'
Problem

We should skip snapshot publishing of a project when the project
version does not end in '-SNAPSHOT'.

Solution

Update the `release_snapshot` function to check that the project
version ends with `-SNAPSHOT` before performing the publish.

Result

The nightly publish does not attempt to publish to snapshots a
non-snapshot version.

Differential Revision: https://phabricator.twitter.biz/D531367