dodo: Ensure we checkout dodo repo for workflows
Problem/Solution

In order to run scripts from the Dodo repo (like the Slack
notification script), the repo needs to be checked out as
part of the workflow. Ensure we checkout the repo for
every workflow that performs Slack notifications additionally
we dump the GitHub context for every workflow which responds
to a repository dispatch event for debugging purposes.

Differential Revision: https://phabricator.twitter.biz/D555455