dodo: Move All Actions From openjdk to adopt
Problem / Solution

Due to recent security related upgrades for GitHub Actions,
'openjdk@1.8' is no longer available as a specific 'java-version'
that can be used. 'adopt@1.8' is a prebuilt version of OpenJDk provided
by AdoptOpenJDK, which is part of the Eclipse Foundation. It's been
proven to work with our 'release-snapshots' action, so now let's roll it
out to all scripts.

JIRA Issues: CSL-10418

Differential Revision: https://phabricator.twitter.biz/D586293